🚚 Sort notion api calls
This commit moves around a lot of code, but there is no change in functionality.
